feat: Add echo command to repeat user-provided text
Added a new functionality to the bot to handle the "/echo" command, allowing it to repeat any text provided by the user. The bot now responds to the "/echo" command by echoing back the text following the command. If no text is provided, the bot replies with a message asking the user to provide text to echo.